backend/notify_weather_forecast/requirements.txt
Flask==2.3.2
pika==1.3.2
python-dotenv==1.0.0
requests==2.31.0